requirements.txt
PyYAML
openai
alfworld[full]